Restructure project according to Spothit new services / Correct http request setup
