+++ src/Exception/RequestException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Partikule\Spothit\Sms\Api\Exception;
+namespace Partikule\Spothit\Exception;
 
 class RequestException extends \RuntimeException
 {
